fix(plan): address all review findings across domain, CLI, and UI layers
Domain:
- normalize CRLF at parser/writer entry points
- escape backslash+newlines in YAML frontmatter strings
- guard nextSubPhaseId overflow past 'z'
- sanitize pipe chars in table row names
- cap filename slug to 60 chars
- Format 2c: skip header-like rows in name column
- Format 4: treat ISO dates as "completed" status
- Format 5: localeCompare tiebreaker for sub-phase sort
- Format 0: force-lowercase phaseId letter suffix
- eliminate double resolvePhaseIds call in scaffoldPlan

CLI:
- split plan-command.ts into read/write handler files (420→182 lines)
- relativize all absolute paths in --json output
- resolvePlanFile walks parent dirs when no target
- validate --priority as P1/P2/P3 before scaffoldPlan

UI:
- add error boundary on KanbanPage phase filtering
- add aria-label on kanban columns, semantic <section> element
- add kanbanLoadError i18n key (EN+VI)
