+++ src/commands/plan/plan-command.ts
@@ -4,17 +4,10 @@
  * Uses ASCII indicators [OK] [!] [X] [i] — no emojis
  */
 import { existsSync, statSync } from "node:fs";
-import { basename, dirname, join, resolve } from "node:path";
-import {
-	buildPlanSummary,
-	parsePlanFile,
-	scanPlanDir,
-	validatePlanFile,
-} from "@/domains/plan-parser/index.js";
-import type { PlanPhase, PlanSummary, ValidationResult } from "@/domains/plan-parser/plan-types.js";
-import { logger } from "@/shared/logger.js";
+import { dirname, join, parse, resolve } from "node:path";
+import type { PlanPhase } from "@/domains/plan-parser/plan-types.js";
 import { output } from "@/shared/output-manager.js";
-import pc from "picocolors";
+import { handleKanban, handleParse, handleStatus, handleValidate } from "./plan-read-handlers.js";
 import { handleAddPhase, handleCheck, handleCreate, handleUncheck } from "./plan-write-handlers.js";
 
 // ─── Options type ─────────────────────────────────────────────────────────────
@@ -40,6 +33,7 @@ export interface PlanCommandOptions {
 /**
  * Resolve a plan file from a target string (file path or directory).
  * If target is a directory, looks for plan.md inside it.
+ * If no target given, walks up parent directories to find plan.md.
  */
 export function resolvePlanFile(target?: string): string | null {
 	const t = target ? resolve(target) : process.cwd();
@@ -52,11 +46,22 @@ export function resolvePlanFile(target?: string): string | null {
 		if (existsSync(candidate)) return candidate;
 	}
 
+	// Walk up parent directories (only when no explicit target)
+	if (!target) {
+		let dir = process.cwd();
+		const root = parse(dir).root;
+		while (dir !== root) {
+			const candidate = join(dir, "plan.md");
+			if (existsSync(candidate)) return candidate;
+			dir = dirname(dir);
+		}
+	}
+
 	return null;
 }
 
 /**
- * Returns true if JSON output is requested via --json flag or --format json
+ * Returns true if JSON output is requested via --json flag
  */
 export function isJsonOutput(options: PlanCommandOptions): boolean {
 	return options.json === true;
@@ -66,7 +71,7 @@ export function isJsonOutput(options: PlanCommandOptions): boolean {
  * Render a simple ASCII progress bar
  * e.g. "[####----]  4/8 (50%)"
  */
-function progressBar(completed: number, total: number, width = 20): string {
+export function progressBar(completed: number, total: number, width = 20): string {
 	if (!Number.isFinite(completed) || !Number.isFinite(total)) return `[${"-".repeat(width)}]  ?/?`;
 	if (total <= 0) return `[${"-".repeat(width)}]  0/0`;
 	const filled = Math.max(0, Math.min(width, Math.round((completed / total) * width)));
@@ -78,7 +83,7 @@ function progressBar(completed: number, total: number, width = 20): string {
 /**
  * Render phases as an ASCII table
  */
-function renderPhasesTable(phases: PlanPhase[]): void {
+export function renderPhasesTable(phases: PlanPhase[]): void {
 	const maxId = Math.max(4, ...phases.map((p) => p.phaseId.length));
 	const maxName = Math.max(4, ...phases.map((p) => p.name.length));
 	const maxStatus = 11; // "in-progress"
@@ -97,249 +102,6 @@ function renderPhasesTable(phases: PlanPhase[]): void {
 		const nameStr = pad(p.name.slice(0, maxName), maxName);
 		console.log(`  ${idStr}  | ${nameStr}  | ${statusIcon} ${p.status}`);
 	}
-}
-
-// ─── Subcommand Handlers ──────────────────────────────────────────────────────
-
-/** parse — output phases as ASCII table or JSON */
-export async function handleParse(
-	target: string | undefined,
-	options: PlanCommandOptions,
-): Promise<void> {
-	const planFile = resolvePlanFile(target);
-	if (!planFile) {
-		output.error(`[X] No plan.md found${target ? ` at '${target}'` : " in current directory"}`);
-		process.exitCode = 1;
-		return;
-	}
-
-	let phases: PlanPhase[];
-	let frontmatter: Record<string, unknown>;
-	try {
-		({ phases, frontmatter } = parsePlanFile(planFile));
-	} catch (err) {
-		output.error(`[X] Failed to read plan: ${err instanceof Error ? err.message : String(err)}`);
-		process.exitCode = 1;
-		return;
-	}
-
-	if (isJsonOutput(options)) {
-		console.log(JSON.stringify({ file: planFile, frontmatter, phases }, null, 2));
-		return;
-	}
-
-	const title =
-		typeof frontmatter.title === "string" ? frontmatter.title : basename(dirname(planFile));
-	console.log();
-	console.log(pc.bold(`  Plan: ${title}`));
-	console.log(`  File: ${planFile}`);
-	console.log(`  Phases found: ${phases.length}`);
-	console.log();
-	if (phases.length > 0) {
-		renderPhasesTable(phases);
-	} else {
-		console.log("  [!] No phases detected");
-	}
-	console.log();
-}
-
-/** validate — format compliance report with line numbers */
-export async function handleValidate(
-	target: string | undefined,
-	options: PlanCommandOptions,
-): Promise<void> {
-	const planFile = resolvePlanFile(target);
-	if (!planFile) {
-		output.error(`[X] No plan.md found${target ? ` at '${target}'` : " in current directory"}`);
-		process.exitCode = 1;
-		return;
-	}
-
-	let result: ValidationResult;
-	try {
-		result = validatePlanFile(planFile, options.strict ?? false);
-	} catch (err) {
-		output.error(`[X] Failed to read plan: ${err instanceof Error ? err.message : String(err)}`);
-		process.exitCode = 1;
-		return;
-	}
-
-	if (isJsonOutput(options)) {
-		console.log(JSON.stringify(result, null, 2));
-		return;
-	}
-
-	console.log();
-	console.log(pc.bold(`  Validating: ${planFile}`));
-	console.log();
-
-	if (result.issues.length === 0) {
-		console.log(`  [OK] No issues found — ${result.phases.length} phases detected`);
-	} else {
-		for (const issue of result.issues) {
-			const icon =
-				issue.severity === "error" ? "[X]" : issue.severity === "warning" ? "[!]" : "[i]";
-			const lineInfo = `L${issue.line}`;
-			console.log(`  ${icon} ${lineInfo}: ${issue.message}  (${issue.code})`);
-			if (issue.fix) console.log(`      Fix: ${issue.fix}`);
-		}
-	}
-
-	console.log();
-	const validStr = result.valid ? pc.green("[OK] Valid") : pc.red("[X] Invalid");
-	console.log(
-		`  ${validStr} — ${result.issues.filter((i) => i.severity === "error").length} errors, ${result.issues.filter((i) => i.severity === "warning").length} warnings`,
-	);
-	console.log();
-
-	if (!result.valid) process.exitCode = 1;
-}
-
-/** status — ASCII progress bar + summary. Lists all plans if given a plans/ dir */
-export async function handleStatus(
-	target: string | undefined,
-	options: PlanCommandOptions,
-): Promise<void> {
-	// Check if target is a plans/ directory (contains plan subdirs, not a plan.md itself)
-	const t = target ? resolve(target) : null;
-	const plansDir =
-		t && existsSync(t) && statSync(t).isDirectory() && !existsSync(join(t, "plan.md")) ? t : null;
-
-	if (plansDir) {
-		// Multi-plan listing mode
-		const planFiles = scanPlanDir(plansDir);
-		if (planFiles.length === 0) {
-			console.log(`  [!] No plans found in ${plansDir}`);
-			return;
-		}
-
-		if (isJsonOutput(options)) {
-			const summaries = planFiles.flatMap((pf) => {
-				try {
-					return [buildPlanSummary(pf)];
-				} catch {
-					return [];
-				}
-			});
-			console.log(JSON.stringify(summaries, null, 2));
-			return;
-		}
-
-		console.log();
-		console.log(pc.bold(`  Plans in: ${plansDir}`));
-		console.log();
-		for (const pf of planFiles) {
-			try {
-				const s = buildPlanSummary(pf);
-				const bar = progressBar(s.completed, s.totalPhases);
-				const title = s.title ?? basename(dirname(pf));
-				console.log(`  ${pc.bold(title)}`);
-				console.log(`  ${bar}`);
-				if (s.inProgress > 0) console.log(`  [~] ${s.inProgress} in progress`);
-				console.log();
-			} catch {
-				console.log(`  [X] Failed to read: ${basename(dirname(pf))}`);
-				console.log();
-			}
-		}
-		return;
-	}
-
-	// Single plan mode
-	const planFile = resolvePlanFile(target);
-	if (!planFile) {
-		output.error(`[X] No plan.md found${target ? ` at '${target}'` : " in current directory"}`);
-		process.exitCode = 1;
-		return;
-	}
-
-	let summary: PlanSummary;
-	try {
-		summary = buildPlanSummary(planFile);
-	} catch (err) {
-		output.error(`[X] Failed to read plan: ${err instanceof Error ? err.message : String(err)}`);
-		process.exitCode = 1;
-		return;
-	}
-
-	if (isJsonOutput(options)) {
-		console.log(JSON.stringify(summary, null, 2));
-		return;
-	}
-
-	const title = summary.title ?? basename(dirname(planFile));
-	console.log();
-	console.log(pc.bold(`  ${title}`));
-	if (summary.status) console.log(`  Status: ${summary.status}`);
-	console.log();
-	console.log(`  Progress: ${progressBar(summary.completed, summary.totalPhases)}`);
-	console.log(`  [OK] Completed:   ${summary.completed}`);
-	console.log(`  [~]  In Progress: ${summary.inProgress}`);
-	console.log(`  [ ]  Pending:     ${summary.pending}`);
-	console.log();
-}
-
-/** kanban — open dashboard at /kanban?file=<path> */
-export async function handleKanban(
-	target: string | undefined,
-	options: PlanCommandOptions,
-): Promise<void> {
-	const planFile = resolvePlanFile(target);
-	if (!planFile) {
-		output.error(`[X] No plan.md found${target ? ` at '${target}'` : " in current directory"}`);
-		process.exitCode = 1;
-		return;
-	}
-
-	logger.info("Starting ClaudeKit Dashboard (Kanban view)...");
-
-	const { port, dev = false } = options;
-	const noOpen = options.open === false;
-
-	let server: { port: number; close: () => Promise<void> };
-	try {
-		const { startServer } = await import("@/domains/web-server/index.js");
-		server = await startServer({ port, openBrowser: false, devMode: dev });
-	} catch (err) {
-		output.error(`[X] Failed to start server: ${err instanceof Error ? err.message : String(err)}`);
-		process.exitCode = 1;
-		return;
-	}
-
-	const encodedPath = encodeURIComponent(planFile);
-	const url = `http://localhost:${server.port}/kanban?file=${encodedPath}`;
-
-	console.log();
-	console.log(pc.bold("  ClaudeKit Dashboard — Kanban"));
-	console.log(pc.dim("  ──────────────────────────────"));
-	console.log(`  Local:  ${pc.cyan(url)}`);
-	console.log(`  File:   ${planFile}`);
-	console.log();
-	console.log(pc.dim("  Press Ctrl+C to stop"));
-	console.log();
-
-	if (!noOpen) {
-		try {
-			const { default: open } = await import("open");
-			await open(url);
-		} catch {
-			// Non-fatal: server still runs, user can open URL manually
-			console.log(pc.dim("  [i] Could not open browser automatically"));
-		}
-	}
-
-	// Block until Ctrl+C or SIGTERM — resolves the promise to let the function return cleanly
-	await new Promise<void>((resolvePromise) => {
-		const shutdown = async () => {
-			console.log();
-			logger.info("Shutting down...");
-			// Race server.close() against a 3s timeout to avoid hanging on open connections
-			await Promise.race([server.close(), new Promise<void>((r) => setTimeout(r, 3000))]);
-			resolvePromise();
-		};
-		process.once("SIGINT", shutdown);
-		process.once("SIGTERM", shutdown);
-	});
 }
 
 // ─── Main dispatcher ──────────────────────────────────────────────────────────
